requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.4
zipp==3.15.0
certifi==2022.12.7
charset-normalizer==3.1.0
DateTime==5.1
elasticsearch==7.17.0
idna==3.4
itsdangerous==2.1.2
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.24.3
pandas==1.4.1
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
tzdata==2023.3
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.4
zope.interface==6.0